requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
accelerate
sentence-transformers
faiss-cpu
numpy
pypdf